Update mdpy.core.environment module, update bit type

+++ mdpy/environment.py
@@ -14,9 +14,9 @@ from mdpy.error import *
 
 precision = 'single'
 
-CUPY_BIT = cp.uint8
-NUMBA_BIT = nb.uint8
-NUMPY_BIT = np.uint8
+CUPY_BIT = cp.uint32
+NUMBA_BIT = nb.uint32
+NUMPY_BIT = np.uint32
 
 if precision == 'single':
     CUPY_FLOAT = cp.float32
